feat(vite, cypress): remove vue-cli, add vite, refactor cypress for testing
We removed vue/cli and webpack completely.  We're using vite to build and run development.  Cypress
now uses vite under the hood and E2E testing and unit testing are functional.  Corrections were made
to the backend and it works now along with the unit tests.  This should be the last large commit
before switching to trunk-based-development.

BREAKING CHANGE: A number of changes

#20